bcmdhd: Disable DHD_SKIP_COREDUMP_ON_HC in current stage
In the Wi-Fi driver/firmware design, if the Wi-Fi dongle's attempt to wake the host system but times out, it triggers a subsystem restart (SSR).
On kernel 5.10, we observed this trap signature was impacting wifi stability, but turns out this signature was actually caused by system performance itself, not wifi issue (b/268578193#comment11), so we decided to ease this wifi trap.
However, on kernel 6.1, the frequency of this SSR increased. To confirm whether this is the same issue or not, a patch was added to generate memory dumps for further debugging.

Bug: 347131644
Test: ACTS test

Change-Id: I7f397e72b79b744c243c3b301d70457722ae8899
Signed-off-by: Dennis Jeon <dennis.jeon@broadcom.corp-partner.google.com>